+++ src/flownet/network_model/_generate_connections.py
@@ -131,7 +131,11 @@ def _split_additional_flow_nodes(
 
     fractions = [v / sum(volumes) for v in volumes]
     sep_add_flownodes = [round(frac * total_additional_nodes) for frac in fractions]
-
+    print(
+        f"The total {str(total_additional_nodes)} additional flow nodes "
+        "are split over the layers based on the volume of the bounding boxes, "
+        f"{str(sep_add_flownodes)}."
+    )
     return sep_add_flownodes
 
 
@@ -405,17 +409,17 @@ def _create_entity_connection_matrix(
             tube_coordinates = linspace(
                 start=start,
                 stop=end,
-                num=n_non_reservoir_evaluation,
+                num=n_non_reservoir_evaluation,  # type: ignore
                 endpoint=False,
                 dtype=float,
                 axis=1,
             ).T
 
             if not any(
-                [
+                (
                     all(check_in_hull(concave_hull, tube_coordinates))
                     for concave_hull in concave_hull_list
-                ]
+                )
             ):
                 continue
 
@@ -590,13 +594,17 @@ def create_connections(
         Desired restructuring of start-end coordinates into separate columns, as per Flow needs.
 
     """
-    if df_coordinates["LAYER_ID"].nunique() > 1 and concave_hull_list is not None:
+    if (
+        df_coordinates["LAYER_ID"].nunique() > 1
+        and concave_hull_list is not None
+        and len(configuration.flownet.additional_flow_nodes) == 1
+    ):
         additional_flow_nodes_list = _split_additional_flow_nodes(
-            total_additional_nodes=configuration.flownet.additional_flow_nodes,
+            total_additional_nodes=configuration.flownet.additional_flow_nodes[0],
             concave_hull_list=concave_hull_list,
         )
     else:
-        additional_flow_nodes_list = [configuration.flownet.additional_flow_nodes]
+        additional_flow_nodes_list = list(configuration.flownet.additional_flow_nodes)
 
     starts: List[Coordinate] = []
     ends: List[Coordinate] = []