+++ src/flownet/parameters/_relative_permeability.py
@@ -1,9 +1,7 @@
 from typing import Dict, List, Tuple, Union, Optional
-import concurrent.futures
-import functools
 
 import pandas as pd
-from pyscal import WaterOilGas, WaterOil, GasOil, PyscalFactory, PyscalList
+import numpy as np
 
 from configsuite import ConfigSuite
 from ..utils import write_grdecl_file
@@ -13,93 +11,150 @@ from .probability_distributions import ProbabilityDistribution
 from ._base_parameter import Parameter, parameter_probability_distribution_class
 
 
-def gen_wog(parameters: pd.DataFrame, fast_pyscal: bool = False) -> WaterOilGas:
+def swof_from_parameters(parameters: Dict) -> str:
     """
-    Creates a PyScal WaterOilGas object based on the input parameters supplied.
+    Creates a SWOF table based on a dictionary of input parameters/values
 
     Args:
-        parameters: A dataframe consisting of all specified parameters.
-        fast_pyscal: Run pyscal in fast-mode skipping checks. Useful for large models/ensemble.
+        parameters: Dictionary of saturation and relative permeability endpoints
 
     Returns:
-        A PyScal WaterOilGas object
-
+        A string with the resulting SWOF table
     """
-    wog_relperm = WaterOilGas(
-        swirr=parameters["swirr"],
-        swl=parameters["swl"],
-        swcr=parameters["swcr"],
-        sorw=parameters["sorw"],
-        sorg=parameters["sorg"],
-        sgcr=parameters["sgcr"],
-        h=H_CONSTANT,
-        fast=fast_pyscal,
-    )
+    swl = parameters["swl"]
+    swcr = parameters["swcr"]
+    sorw = parameters["sorw"]
+    kroend = parameters["kroend"]
+    krwend = parameters["krwend"]
+    now = parameters["now"]
+    nw = parameters["nw"]
 
-    wog_relperm.wateroil.add_corey_water(
-        nw=parameters["nw"], krwend=parameters["krwend"]
-    )
-    wog_relperm.wateroil.add_corey_oil(
-        now=parameters["now"], kroend=parameters["kroend"]
-    )
-    wog_relperm.gasoil.add_corey_gas(ng=parameters["ng"], krgend=parameters["krgend"])
-    wog_relperm.gasoil.add_corey_oil(nog=parameters["nog"], kroend=parameters["kroend"])
+    # array of water saturations to calculate krow and krw for
+    sw = np.sort(np.append(np.arange(swl, 1, H_CONSTANT), [swcr, 1 - sorw, 1]))
+    # remove potential duplicate values
+    sw = sw[~(np.triu(np.abs(sw[:, None] - sw) <= 1e-5, 1)).any(0)]
+    # normalized saturations
+    swn = ((sw - swcr) / (1 - swcr - sorw)).clip(min=0)
+    son = ((1 - sw - sorw) / (1 - sorw - swl)).clip(min=0)
+    # calculate relative permeabilities
+    krow = kroend * son ** now
+    krw = krwend * swn ** nw
+    # interpolate between krwend and krwmax (=1)
+    krw_interp = (1 - krwend) / sorw * (sw - 1) + 1
+    krw[krw > krwend] = krw_interp[krw > krwend]
+    # only zero capillary pressure implemented
+    pc = np.zeros(len(sw))
 
-    return wog_relperm
+    swof = np.transpose(np.stack((sw, krw, krow, pc)))
+    swof_string = np.array2string(swof, formatter={"float_kind": lambda x: "%.7f" % x})
+    swof_string = swof_string.replace(" [", "").replace("[", "").replace("]", "")
+
+    return swof_string
 
 
-def gen_wo(parameters: pd.DataFrame, fast_pyscal: bool = False) -> WaterOil:
+def sgof_from_parameters(parameters: Dict) -> str:
     """
-    Creates a PyScal WaterOil object based on the input parameters supplied.
+    Creates a SGOF table based on a dictionary of input parameters/values
 
     Args:
-        parameters: A dataframe consisting of all specified parameters.
-        fast_pyscal: Run pyscal in fast-mode skipping checks. Useful for large models/ensembles.
+        parameters: Dictionary of saturation and relative permeability endpoints
 
     Returns:
-        A PyScal WaterOil object
-
+        A string with the resulting SGOF table
     """
-    wo_relperm = WaterOil(
-        swirr=parameters["swirr"],
-        swl=parameters["swl"],
-        swcr=parameters["swcr"],
-        sorw=parameters["sorw"],
-        h=H_CONSTANT,
-        fast=fast_pyscal,
+    swl = parameters["swl"]
+    sorg = parameters["sorg"]
+    sgcr = parameters["sgcr"]
+    krgend = parameters["krgend"]
+    kroend = parameters["kroend"]
+    ng = parameters["ng"]
+    nog = parameters["nog"]
+
+    # array of gas saturations to calculate krog and krg for
+    sg = np.sort(
+        np.append(np.arange(sgcr, 1 - swl, H_CONSTANT), [1 - sorg - swl, 1 - swl, 0])
     )
+    # remove potential duplicate values
+    sg = sg[~(np.triu(np.abs(sg[:, None] - sg) <= 1e-5, 1)).any(0)]
+    # normalized saturations
+    son = (((1 - sg) - sorg - swl) / (1 - sorg - swl)).clip(min=0)
+    sgn = ((sg - sgcr) / (1 - swl - sgcr - sorg)).clip(min=0)
+    # calculate relative permeabilities
+    krog = kroend * son ** nog
+    krg = krgend * sgn ** ng
+    # interpolate between krgend and krgmax (=1)
+    krg_interp = (1 - krgend) / sorg * (sg - 1) + 1 + swl / sorg * (1 - krgend)
+    krg[krg > krgend] = krg_interp[krg > krgend]
+    # only zero capillary pressure implemented
+    pc = np.zeros(len(sg))
 
-    wo_relperm.add_corey_water(nw=parameters["nw"], krwend=parameters["krwend"])
-    wo_relperm.add_corey_oil(now=parameters["now"], kroend=parameters["kroend"])
+    sgof = np.transpose(np.stack((sg, krg, krog, pc)))
+    sgof_string = np.array2string(sgof, formatter={"float_kind": lambda x: "%.7f" % x})
+    sgof_string = sgof_string.replace(" [", "").replace("[", "").replace("]", "")
 
-    return wo_relperm
+    return sgof_string
 
 
-def gen_og(parameters: pd.DataFrame, fast_pyscal: bool = False) -> GasOil:
+def interpolate_wo(parameter: float, scalrec: Dict) -> Dict:
     """
-    Creates a PyScal GasOil object based on the input parameters supplied.
+    Creates interpolated saturation endpoints and relative
+    permeability endpoints for water/oil based on an interpolation parameter
+    and three separate input cases (low/base/high).
 
     Args:
-        parameters: A dataframe consisting of all specified parameters.
-        fast_pyscal: Run pyscal in fast-mode skipping checks. Useful for large models/ensembles.
+        parameter: A value on the interval -1 to 1 used for interpolation
+        scalrec: A dictionary containing the relative permeability
+            and saturation endpoints for the low/base/high cases
 
     Returns:
-        A PyScal GasOil object
-
+        A dictionary with the interpolated water/oil saturation and relative
+        permeability endpoints
     """
-    og_relperm = GasOil(
-        swirr=parameters["swirr"],
-        swl=parameters["swl"],
-        sorg=parameters["sorg"],
-        sgcr=parameters["sgcr"],
-        h=H_CONSTANT,
-        fast=fast_pyscal,
-    )
+    # interpolate swirr, swl, swcr, sorw, nw, now, krwend, kroend
+    if parameter < 0:
+        (i, j) = (1, 0)
+    else:
+        (i, j) = (1, 2)
+    parameter = abs(parameter)
 
-    og_relperm.add_corey_gas(ng=parameters["ng"], krgend=parameters["krgend"])
-    og_relperm.add_corey_oil(nog=parameters["nog"], kroend=parameters["kroend"])
+    parameter_dict = {}
+    for elem in {"swirr", "swl", "swcr", "sorw", "nw", "now", "krwend", "kroend"}:
+        parameter_dict[elem] = (
+            scalrec[elem][i] * (1 - parameter) + scalrec[elem][j] * parameter
+        )
 
-    return og_relperm
+    return parameter_dict
+
+
+def interpolate_go(parameter: float, scalrec: Dict) -> Dict:
+    """
+    Creates interpolated saturation endpoints and relative
+    permeability endpoints for gas/oil based on an interpolation parameter
+    and three separate input cases (low/base/high).
+
+    Args:
+        parameter: A value on the interval -1 to 1 used for interpolation
+        scalrec: A dataframe containing the relative permeability
+            and saturation endpoints for the low/base/high cases
+
+    Returns:
+        A dictionary with the interpolated gas/oil saturation and relative
+        permeability endpoints
+    """
+    # interpolate swirr, swl, sgcr, sorg, ng, nog, krgend, kroend
+    if parameter < 0:
+        (i, j) = (1, 0)
+    else:
+        (i, j) = (1, 2)
+    parameter = abs(parameter)
+
+    parameter_dict = {}
+    for elem in {"swirr", "swl", "sgcr", "sorg", "ng", "nog", "krgend", "kroend"}:
+        parameter_dict[elem] = (
+            scalrec[elem][i] * (1 - parameter) + scalrec[elem][j] * parameter
+        )
+
+    return parameter_dict
 
 
 class RelativePermeability(Parameter):
@@ -154,13 +209,9 @@ class RelativePermeability(Parameter):
         self._phases = config.flownet.phases
         self._fast_pyscal: bool = config.flownet.fast_pyscal
         self._interpolation_values: Optional[pd.DataFrame] = None
-        self._scal_for_interp: Optional[PyscalList] = None
+
         if isinstance(interpolation_values, pd.DataFrame):
             self._interpolation_values = interpolation_values
-            # self._scal_for_interp will be indexed by SATNUM region, starting at 1
-            self._scal_for_interp = PyscalFactory.create_scal_recommendation_list(
-                interpolation_values, fast=self._fast_pyscal
-            )
 
         self._swof, self._sgof = self._check_parameters()
         self._independent_interpolation = (
@@ -272,49 +323,49 @@ class RelativePermeability(Parameter):
             )
             parameters.append(param_value_dict)
 
-        partial_gen_wog = functools.partial(gen_wog, fast_pyscal=True)
-        partial_gen_wo = functools.partial(gen_wo, fast_pyscal=True)
-        partial_gen_og = functools.partial(gen_og, fast_pyscal=True)
-
-        if self._scal_for_interp is not None:
-            wog_list = self._scal_for_interp.interpolate(
-                [param.get("interpolate") for param in parameters],
-                [param.get("interpolate gas") for param in parameters]
-                if self._independent_interpolation
-                else None,
-                h=H_CONSTANT,
-            )
-            str_props_section = wog_list.SWOF()
-            str_props_section += wog_list.SGOF()
-        else:
-            if self._swof and self._sgof:
-                with concurrent.futures.ProcessPoolExecutor() as executor:
-                    for _, relperm in zip(  # type: ignore[assignment]
-                        parameters, executor.map(partial_gen_wog, parameters)
-                    ):
-                        str_swofs += relperm.SWOF(header=False)
-                        str_sgofs += relperm.SGOF(header=False)
-            elif self._swof:
-                with concurrent.futures.ProcessPoolExecutor() as executor:
-                    for _, relperm in zip(  # type: ignore[assignment]
-                        parameters, executor.map(partial_gen_wo, parameters)
-                    ):
-                        str_swofs += relperm.SWOF(header=False)
-            elif self._sgof:
-                with concurrent.futures.ProcessPoolExecutor() as executor:
-                    for _, relperm in zip(  # type: ignore[assignment]
-                        parameters, executor.map(partial_gen_og, parameters)
-                    ):
-                        str_sgofs += relperm.SGOF(header=False)
-            else:
-                raise ValueError(
-                    "It seems like both SWOF and SGOF should not be generated."
-                    "Either one of the two should be generated. Can't continue..."
+        if self._interpolation_values is not None:
+            for satnum in self._unique_satnums:
+                # all endpoints and corey exponents are interpolated
+                # linearly between two of the input case - between
+                # 'low' and  'base' if parameter < 0
+                # and between 'base' and 'high' if parameter >= 0.
+                # pyscal interpolates endpoints and then interpolates
+                # between the non-linear parts of the relative permeability
+                # curves wrt normalised saturation (not interpolating corey
+                # exponents - using scipy interp1d). Resulting SWOFs and
+                # SGOFs are seemingly pretty similar, but not equal.
+                parameter_dict = (
+                    self._interpolation_values.loc[
+                        self._interpolation_values["SATNUM"] == satnum
+                    ]
+                    .reset_index(drop=True)
+                    .to_dict()
+                )
+                interp_wo = parameters[satnum - 1].get("interpolate")
+                interp_go = (
+                    parameters[satnum - 1].get("interpolate gas")
+                    if self._independent_interpolation
+                    else interp_wo
                 )
 
-            str_props_section = f"SWOF\n{str_swofs}\n" if self._swof else ""
-            str_props_section += f"SGOF\n{str_sgofs}\n" if self._sgof else ""
-
+                if isinstance(interp_wo, float) and isinstance(interp_go, float):
+                    param_wo = interpolate_wo(interp_wo, parameter_dict)
+                    param_go = interpolate_go(interp_go, parameter_dict)
+                else:
+                    raise ValueError(
+                        "Interpolation parameter is not a float."
+                        "Something wrong with parameters generated by ERT"
+                    )
+                str_swofs += swof_from_parameters(param_wo) + "\n/\n"
+                str_sgofs += sgof_from_parameters(param_go) + "\n/\n"
+        else:
+            for param in parameters:
+                if self._swof:
+                    str_swofs += swof_from_parameters(param) + "\n/\n"
+                if self._sgof:
+                    str_sgofs += sgof_from_parameters(param) + "\n/\n"
+        str_props_section = f"SWOF\n{str_swofs}\n" if self._swof else ""
+        str_props_section += f"SGOF\n{str_sgofs}\n" if self._sgof else ""
         str_runspec_section = "\n".join(self._phases).upper() + "\n"
 
         return {